requirements.txt
beautifulsoup4>=4.9.0
requests>=2.25.0
selenium>=4.0.0
tldextract>=3.1.0
openpyxl>=3.0.0
PyYAML>=5.4.0
regex>=2021.4.0
aiohttp>=3.8.0
pygments>=2.10.0
